Remove user job

+++ app/Livewire/LikeButton.php
@@ -3,6 +3,7 @@
 namespace App\Livewire;
 
 use App\Models\Likes;
+use App\Models\User;
 use Illuminate\Support\Facades\Auth;
 use Livewire\Component;
 
@@ -20,23 +21,26 @@ class LikeButton extends Component
 
     public function toggleLike()
     {
-        $existingLike = Likes::where('liker_id', Auth::id())
-            ->where('liked_user_id', $this->userId)
-            ->first();
+        if ($this->checkRole()) {
+            $existingLike = Likes::where('liker_id', Auth::id())
+                ->where('liked_user_id', $this->userId)
+                ->first();
 
-        if ($existingLike) {
-            // If a like exists, remove it
-            $existingLike->delete();
-            $this->liked = false;
-        } else {
-            // If a like doesn't exist, create it
-            Likes::create([
-                'liker_id' => Auth::id(),
-                'liked_user_id' => $this->userId
-            ]);
-            $this->liked = true;
+            if ($existingLike) {
+                // If a like exists, remove it
+                $existingLike->delete();
+                $this->liked = false;
+            } else {
+                // If a like doesn't exist, create it
+                Likes::create([
+                    'liker_id' => Auth::id(),
+                    'liked_user_id' => $this->userId
+                ]);
+                $this->liked = true;
+            }
         }
     }
+
 
     public function checkIfLiked()
     {
@@ -50,5 +54,17 @@ class LikeButton extends Component
     public function render()
     {
         return view('livewire.like-button');
+    }
+
+    public function checkRole()
+    {
+        $liker = Auth::user();
+        $liked = User::find($this->userId);
+
+        if ($liker->role == $liked->role) {
+            return false;
+        } else {
+            return true;
+        }
     }
 }
